+++ ninaAPI/WebService/V2/ControllerV2.cs
@@ -9,6 +9,8 @@
 
 #endregion "copyright"
 
+using System;
+using System.Globalization;
 using System.Reflection;
 using EmbedIO;
 using EmbedIO.Routing;
@@ -29,6 +31,24 @@ namespace ninaAPI.WebService.V2
         public void GetVersion()
         {
             HttpContext.WriteToResponse(new HttpResponse() { Response = Assembly.GetAssembly(typeof(AdvancedAPI)).GetName().Version.ToString() });
+        }
+
+        [Route(HttpVerbs.Get, "/time")]
+        public void GetTime()
+        {
+            HttpContext.WriteToResponse(new HttpResponse() { Response = DateTime.Now });
+        }
+
+        [Route(HttpVerbs.Get, "/application-start")]
+        public void GetApplicationStart()
+        {
+            HttpContext.WriteToResponse(new HttpResponse() { Response = NINA.Core.Utility.CoreUtil.ApplicationStartDate });
+        }
+
+        [Route(HttpVerbs.Get, "/version/nina")]
+        public void GetNINAVersion([QueryField] bool friendly)
+        {
+            HttpContext.WriteToResponse(new HttpResponse() { Response = friendly ? NINA.Core.Utility.CoreUtil.VersionFriendlyName : NINA.Core.Utility.CoreUtil.Version });
         }
     }
 }